feat: :lock: auth in controller

+++ users/src/infraestructure/controller/user.controller.ts
@@ -1,4 +1,5 @@
 import {
+  Headers,
   Body,
   ConflictException,
   Controller,
@@ -9,39 +10,46 @@ import {
   Patch,
   Post,
   Get,
+  UnauthorizedException,
 } from '@nestjs/common';
-import { ApiTags, ApiOperation, ApiBody, ApiOkResponse } from '@nestjs/swagger';
+import {
+  ApiTags,
+  ApiOperation,
+  ApiBody,
+  ApiOkResponse,
+  ApiBearerAuth,
+} from '@nestjs/swagger';
 
 import {
   UserAlreadyExistsError,
   UserNotFoundError,
 } from '../../domain/user/exception';
-import { USER_SERVICE, UserService } from '../service';
+import { AuthService, USER_SERVICE, UserService } from '../service';
 import { LoginUserDTO, UserDTO } from 'src/utils';
 import { LoginUserResponse } from 'src/utils/user/LoginUserResponse';
+import { ReadAllResponse, UserProps } from 'src/utils/user/ReadAllResponse';
+import { AUTH_REPOSITORY } from 'src/domain';
 
 @ApiTags('UserController')
 @Controller('user')
 export class UserController {
   constructor(
     @Inject(USER_SERVICE) private readonly userService: UserService,
+    @Inject(AUTH_REPOSITORY) private readonly authService: AuthService,
   ) {}
-
-  /*
-  @Get()
-  @ApiOperation({ summary: 'Obtener todos los usuarios' })
-  @ApiOkResponse({ type: UserDTO })
-  async findAll(): Promise<UserDTO[] | null> {
-    return null;
-    
-  }
-
-  */
 
   @Get(':id')
   @ApiOperation({ summary: 'Obtener un usuario por id' })
   @ApiOkResponse({ type: UserDTO })
-  async findOne(@Param('id') id: string): Promise<UserDTO | null> {
+  @ApiBearerAuth()
+  async findOne(
+    @Param('id') id: string,
+    @Headers('authorization') jwt: string,
+  ): Promise<UserDTO | null> {
+    const isTokenValid = await this.authService.verifyToken(jwt);
+    if (!isTokenValid) {
+      throw new UnauthorizedException('Invalid token');
+    }
     try {
       const user: UserDTO = await this.userService.readUser(id);
       return user;
@@ -58,7 +66,15 @@ export class UserController {
   @ApiOperation({ summary: 'Crear un usuario' })
   @ApiBody({ type: UserDTO })
   @ApiOkResponse({ type: UserDTO })
-  async create(@Body() userDto: UserDTO): Promise<void> {
+  @ApiBearerAuth()
+  async create(
+    @Body() userDto: UserDTO,
+    @Headers('authorization') jwt: string,
+  ): Promise<void> {
+    const isTokenValid = await this.authService.verifyToken(jwt);
+    if (!isTokenValid) {
+      throw new UnauthorizedException('Invalid token');
+    }
     try {
       await this.userService.createUser(userDto);
       return;
@@ -75,7 +91,15 @@ export class UserController {
   @ApiOperation({ summary: 'Login de un usuario' })
   @ApiBody({ type: LoginUserDTO })
   @ApiOkResponse({ type: LoginUserDTO })
-  async login(@Body() userDto: LoginUserDTO): Promise<LoginUserResponse> {
+  @ApiBearerAuth()
+  async login(
+    @Body() userDto: LoginUserDTO,
+    @Headers('authorization') jwt: string,
+  ): Promise<LoginUserResponse> {
+    const isTokenValid = await this.authService.verifyToken(jwt);
+    if (!isTokenValid) {
+      throw new UnauthorizedException('Invalid token');
+    }
     try {
       return await this.userService.loginUser(userDto);
     } catch (e) {
@@ -91,7 +115,15 @@ export class UserController {
   @ApiOperation({ summary: 'Actualizar un usuario' })
   @ApiBody({ type: UserDTO })
   @ApiOkResponse({ type: UserDTO })
-  async update(@Body() userDto: UserDTO): Promise<UserDTO> {
+  @ApiBearerAuth()
+  async update(
+    @Body() userDto: UserDTO,
+    @Headers('authorization') jwt: string,
+  ): Promise<UserDTO> {
+    const isTokenValid = await this.authService.verifyToken(jwt);
+    if (!isTokenValid) {
+      throw new UnauthorizedException('Invalid token');
+    }
     try {
       return await this.userService.updateUser(userDto);
     } catch (e) {
@@ -106,7 +138,15 @@ export class UserController {
   @Delete(':id')
   @ApiOperation({ summary: 'Eliminar un usuario por id' })
   @ApiOkResponse()
-  async delete(@Param('id') id: string): Promise<string> {
+  @ApiBearerAuth()
+  async delete(
+    @Param('id') id: string,
+    @Headers('authorization') jwt: string,
+  ): Promise<string> {
+    const isTokenValid = await this.authService.verifyToken(jwt);
+    if (!isTokenValid) {
+      throw new UnauthorizedException('Invalid token');
+    }
     try {
       await this.userService.deleteUser(id);
       return 'User <' + id + '> deleted ';
@@ -122,82 +162,18 @@ export class UserController {
   @Get()
   @ApiOperation({ summary: 'Obtener todos los usuarios' })
   @ApiOkResponse({ type: UserDTO })
-  async findAll(): Promise<UserDTO[] | null> {
-    return;
-  }
-  /*
-  @Post('login')
-  async login(@Body() loginDTO: LoginDTO): Promise<TokenResponse> {
-    const { email, password } = loginDTO;
-
-    const user: UserDTO = await this.userService.getUserByEmail(email);
-
-    if (!user) {
-      throw new UnauthorizedException('Login failed');
-    }
-
-    const isValidPassword = await this.authService.validatePassword(
-      password,
-      user.password
-    );
-
-    if (!isValidPassword) {
-      throw new UnauthorizedException('Login failed');
-    }
-
-    return new TokenResponse(await this.authService.generateToken(user.id));
-  }
-
-  @Post('validate-token')
-  @ApiOperation({ summary: 'Validar el token del usuario' })
-  @ApiBody({ type: TokenResponse })
-  async validateToken(@Body() token: TokenResponse): Promise<RoleResponse> {
-    const userId: UserId = await this.authService.validateToken(token.token);
-
-    if (!userId) {
+  @ApiBearerAuth()
+  async findAll(
+    @Headers('authorization') jwt: string,
+  ): Promise<ReadAllResponse> {
+    const isTokenValid = await this.authService.verifyToken(jwt);
+    if (!isTokenValid) {
       throw new UnauthorizedException('Invalid token');
     }
-
-    return new RoleResponse(
-      (await this.userService.getUserById(userId.value)).role,
-      userId.value
-    );
-  }
-
-  @Post('change-password')
-  @ApiOperation({ summary: 'Cambiar la contraseña del usuario' })
-  @ApiOkResponse({ type: ChangePasswordRequestDTO })
-  async changePassword(
-    @Body() changePasswordDto: ChangePasswordRequestDTO
-  ): Promise<ChangePasswordResponse> {
-    const { token, password, newpassword } = changePasswordDto;
-    try {
-      const userId: UserId = await this.authService.validateToken(token);
-      const user: UserDTO = await this.userService.getUserById(userId.value);
-      const isValidPassword = await this.authService.validatePassword(
-        password,
-        user.password
-      );
-
-      if (!isValidPassword) {
-        throw new ForbiddenException('Contraseña incorrecta');
-      }
-
-      const newHashedPassword = await this.authService.hashPassword(
-        newpassword
-      );
-      await this.userService.updateUserPassword(userId, newHashedPassword);
-    } catch (e) {
-      if (e instanceof ForbiddenException) {
-        throw e;
-      } else {
-        throw new UnauthorizedException(e.message);
-      }
-    }
+    const users: UserProps[] = await this.userService.readAllUsers();
     return {
-      token: token,
-      message: 'Contraseña cambiada correctamente'
+      jwt,
+      users,
     };
   }
-  */
 }
